// src/api/model/invoiceListByCompanyKind.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */

export type InvoiceListByCompanyKind = typeof InvoiceListByCompanyKind[keyof typeof InvoiceListByCompanyKind];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceListByCompanyKind = {
  INVOICE: 'INVOICE',
  SIMPLE: 'SIMPLE',
} as const;